ta1394/general: split crates for AV/C operation defined in general specification
This commit adds "ta1394-avc-general" crate. The crates includes a trait
to perform AV/C operations defined in below documentation.

 * AV/C Digital Interface Command Set General Specification Version 4.2
   (Sep. 1, 2004, 1394 Trading Association, Document number 2004006)

The crate also includes some structures as implementation of some AV/C
commands defined in the above document.

Additionally, the crate includes trait and its implementation to parse
content of Configuration ROM for standard AV/C device defined in below
documentation.

 * Configuration ROM for AV/C Devices 1.0 (Dec. 12, 2000, 1394 Trading
   Association, TA Document 1999027)

The parser implementation is for ConfigRom structure provided by
ieee1212-config-rom trait.

 * https://crates.io/crates/ieee1212-config-rom
