+++ Cargo.toml
@@ -31,6 +31,7 @@ members = [
     "libs/alsa-ctl-tlv-codec",
     "libs/core",
     "libs/ieee1212-config-rom",
+    "libs/ta1394/general",
     "libs/ta1394",
     "libs/dg00x/protocols",
     "libs/dg00x/runtime",
@@ -54,6 +55,7 @@ members = [
 [patch.crates-io]
 alsa-ctl-tlv-codec = { path = "libs/alsa-ctl-tlv-codec" }
 ieee1212-config-rom = { path = "libs/ieee1212-config-rom" }
+ta1394-avc-general = { path = "libs/ta1394/general" }
 ta1394 = { path = "libs/ta1394" }
 bebob-protocols = { path = "libs/bebob/protocols" }
 dg00x-protocols = { path = "libs/dg00x/protocols" }
