+++ libs/bebob/runtime/src/lib.rs
@@ -32,7 +32,7 @@ use {
     model::*,
     nix::sys::signal,
     std::{convert::TryFrom, sync::mpsc},
-    ta1394::config_rom::*,
+    ta1394_avc_general::config_rom::*,
 };
 
 enum Event {
