+++ libs/dg00x/runtime/Cargo.toml
@@ -18,6 +18,6 @@ hitaki = "0.2"
 alsactl = "0.4"
 alsa-ctl-tlv-codec = "0.1"
 ieee1212-config-rom = "0.1"
-ta1394 = "0.1"
+ta1394-avc-general = "0.1"
 dg00x-protocols = "0.1"
 core = { path = "../../core" }
