+++ libs/oxfw/protocols/Cargo.toml
@@ -14,4 +14,5 @@ repository = "https://github.com/alsa-project/snd-firewire-ctl-services"
 [dependencies]
 glib = "0.15"
 hinawa = "0.7"
+ta1394-avc-general = "0.1"
 ta1394 = "0.1"
