+++ libs/oxfw/runtime/Cargo.toml
@@ -17,6 +17,7 @@ hinawa = "0.7"
 hitaki = "0.2"
 alsactl = "0.4"
 ieee1212-config-rom = "0.1"
+ta1394-avc-general = "0.1"
 ta1394 = "0.1"
 oxfw-protocols = "0.1"
 core = { path = "../../core" }
