+++ libs/oxfw/runtime/src/common_ctl.rs
@@ -4,7 +4,8 @@
 use {
     super::*,
     std::marker::PhantomData,
-    ta1394::{amdtp::*, general::*, stream_format::*, *},
+    ta1394_avc_general::{general::*, *},
+    ta1394::{amdtp::*, stream_format::*},
 };
 
 #[derive(Default, Debug)]
